Use explicit typecasts to force signed char comparisons
Avoids warnings about tautological comparisons (either the >= 0 comparison for ARM/PowerPC or the <= 127 comparison for x86*)

+++ Source/utils/utf8.cpp
@@ -28,7 +28,7 @@ string_view TruncateUtf8(string_view str, std::size_t len)
 char32_t DecodeFirstUtf8CodePoint(string_view input, std::size_t *len)
 {
 	uint32_t codepoint = 0;
-	uint32_t state = UTF8_ACCEPT;
+	uint8_t state = UTF8_ACCEPT;
 	for (std::size_t i = 0; i < input.size(); ++i) {
 		state = utf8_decode_step(state, static_cast<uint8_t>(input[i]), &codepoint);
 		if (state == UTF8_ACCEPT) {
